requirements.txt
fastapi~=0.115.12
uvicorn
hypercorn
requests
python-dotenv~=1.1.0
google-genai
pydantic~=2.11.4
psycopg2~=2.9.10
protobuf~=5.29.4
python-multipart

Markdown~=3.8
beautifulsoup4~=4.13.4
supabase~=2.15.1